requirements.txt
asgiref==3.5.2
Django==4.1.1 
django-filter==22.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
Pillow==9.2.0
psycopg2-binary==2.9.5
pytz==2022.2.1
sqlparse==0.4.3